+++ src/test/java/kitchenpos/product/service/ProductServiceTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.product;
+package kitchenpos.product.service;
 
 import kitchenpos.application.ProductService;
 import kitchenpos.domain.Product;
@@ -6,16 +6,12 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.web.util.NestedServletException;
 
 import java.math.BigDecimal;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.hamcrest.Matchers.hasSize;
 import static org.junit.jupiter.api.Assertions.assertThrows;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 @SpringBootTest
 public class ProductServiceTest {
