+++ src/test/java/kitchenpos/table/controller/TableControllerTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.table;
+package kitchenpos.table.controller;
 
 import kitchenpos.common.ControllerTest;
 import kitchenpos.domain.Order;
